+++ kernel/world/object.cpp
@@ -3,8 +3,8 @@
 namespace detail {
 namespace world {
 
-void LocalObject::Normalize(double scale) {
-  double max = -1.0;
+void LocalObject::Normalize(float scale) {
+  float max = -1.0;
   for (const geometry::Point &point : vertexes) {
     max = std::max(max, abs(point.x));
     max = std::max(max, abs(point.y));
@@ -18,10 +18,13 @@ void LocalObject::Normalize(double scale) {
 }
 
 geometry::Triangle LocalObject::GetTriangle(const TriangleKeeper &triangle) const {
-  return geometry::Triangle{vertexes[triangle.first_index], vertexes[triangle.second_index], vertexes[triangle.third_index]};
+  return geometry::Triangle{vertexes[triangle.first_index], vertexes[triangle.second_index],
+                            vertexes[triangle.third_index]};
 }
 
-geometry::Segment LocalObject::GetSegment(const SegmentKeeper &keeper) const { return geometry::Segment{vertexes[keeper.first_index], vertexes[keeper.second_index]}; }
+geometry::Segment LocalObject::GetSegment(const SegmentKeeper &keeper) const {
+  return geometry::Segment{vertexes[keeper.first_index], vertexes[keeper.second_index]};
+}
 
 const LocalObject::Vertexes &LocalObject::GetVertexes() const { return vertexes; }
 
@@ -31,11 +34,17 @@ const LocalObject::Segments &LocalObject::GetSegments() const { return segments;
 
 LocalObjectBuilder::LocalObjectBuilder() : local_object() {};
 
-void LocalObjectBuilder::AddVertex(const geometry::Point &point) { local_object.vertexes.push_back(point); }
+void LocalObjectBuilder::AddVertex(const geometry::Point &point) {
+  local_object.vertexes.push_back(point);
+}
 
-void LocalObjectBuilder::AddTriangle(const TriangleKeeper &triangle) { local_object.triangles.push_back(triangle); }
+void LocalObjectBuilder::AddTriangle(const TriangleKeeper &triangle) {
+  local_object.triangles.push_back(triangle);
+}
 
-void LocalObjectBuilder::AddSegment(const SegmentKeeper &segment) { local_object.segments.push_back(segment); }
+void LocalObjectBuilder::AddSegment(const SegmentKeeper &segment) {
+  local_object.segments.push_back(segment);
+}
 
 LocalObject LocalObjectBuilder::Extract() { return std::move(local_object); }
